pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ghost-sentry"
version = "0.1.0"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "textual>=0.47.0",
    "rasterio>=1.3.0",
    "shapely>=2.0.0",
    "ultralytics>=8.1.0",
    "pydantic>=2.5.0",
    "numpy",
    "scipy",
    "httpx",
    "python-dotenv",
    "websockets>=12.0",
]

[project.optional-dependencies]
dev = ["pytest", "black", "isort", "mypy"]

[tool.hatch.build.targets.wheel]
packages = ["src/ghost_sentry"]
